+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## 17.3.0 (upcoming)
 
+**New features:**
+ 
+* st-input: Allow to display an autocomplete list in order to select a predefined text
+
 **Others:**
 
 * Update icon font to 1.26